requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database (PostgreSQL)
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# AI and LLM
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1

# Vector operations and similarity
numpy==1.24.3
faiss-cpu==1.7.4
scikit-learn==1.3.2

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
unstructured==0.11.8
markdown==3.5.1

# OCR (Optical Character Recognition)
pytesseract==0.3.10
Pillow==10.0.0
pdf2image==1.16.3
opencv-python==4.8.1.78

# Text processing
tiktoken==0.5.2
langchain==0.1.0
langchain-text-splitters==0.0.1

# Web and HTTP
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2

# Data validation and settings
pydantic==2.5.2
pydantic-settings==2.1.0

# Utilities
python-magic==0.4.27
chardet==5.2.0
aiofiles==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Background tasks
celery==5.3.4
redis==5.0.1